the chart ticks are now formated
