+++ src/utils/parseChartData.ts
@@ -1,3 +1,5 @@
+import { formatChartDate } from "./formatUtils";
+import { getDateKeyForInterval } from "./getDateKeyForInterval";
 import { ChartData, ParsedChartData } from "./types/ChartData";
 
 export const parseChartData = (
@@ -15,12 +17,12 @@ export const parseChartData = (
 
   if (reducedChartData.length === 1 && reducedChartData[0])
     reducedChartData[0].forEach((value, key) =>
-      result.push({ time: key, [coinName1]: value }),
+      result.push({ time: formatChartDate(key), [coinName1]: value }),
     );
   else
     reducedChartData[0].forEach((value, key) =>
       result.push({
-        time: key,
+        time: formatChartDate(key),
         [coinName1]: value,
         [coinName2]: reducedChartData[1].get(key),
       }),
@@ -56,32 +58,6 @@ const reduceChartData = (
   key: "prices" | "market_caps" | "total_volumes",
   interval: string,
 ) => {
-  const getDateKeyForInterval = (
-    timestamp: number,
-    interval: string,
-  ): string => {
-    const date = new Date(timestamp);
-
-    switch (interval) {
-      case "1D":
-        return `${date.getUTCHours()}:00`;
-      case "7D":
-      case "14D": {
-        const hours = date.getUTCHours();
-        const roundedHours =
-          interval === "7D"
-            ? Math.floor(hours / 4) * 4
-            : Math.floor(hours / 8) * 8;
-        return `${date.getUTCFullYear()}-${date.getUTCMonth()}-${date.getUTCDate()} ${roundedHours}:00`;
-      }
-      case "1M":
-      case "1Y":
-        return `${date.getUTCFullYear()}-${date.getUTCMonth()}-${date.getUTCDate()}`;
-      default:
-        return `${date.getUTCFullYear()}-${date.getUTCMonth()}-${date.getUTCDate()} ${date.getUTCHours()}:00`;
-    }
-  };
-
   const reduceCallback = (acc: Map<string, number>, cur: number[]) => {
     const [timestamp, value] = cur;
     const dateKey = getDateKeyForInterval(timestamp, interval);
